requirements.txt
matplotlib==3.8.2
opencv-python==4.8.1.78
keras==2.15.0
tensorflow==2.15.0
gradio==4.7.1
Pillow==10.1.0
sklearn
